+++ src/count-processing-service/domain/repository.go
@@ -10,4 +10,7 @@ type CountValueRepository interface {
 	GetByID(ctx context.Context, itemID string) (*CountValue, error)
 	GetByIDs(ctx context.Context, itemIDs []string) ([]CountValue, error)
 	Delete(ctx context.Context, itemID string) error
+	Increase(ctx context.Context, itemID string, amount int) (*CountValue, error)
+	Decrease(ctx context.Context, itemID string, amount int) (*CountValue, error)
+	Reset(ctx context.Context, itemID string) (*CountValue, error)
 }